Fix payment amount type
Make it float rather than int

+++ src/payment/CashOnDeliveryPayment.java
@@ -2,7 +2,7 @@ package payment;
 
 public class CashOnDeliveryPayment implements PaymentMethod {
     @Override
-    public void pay(int amount) {
+    public void pay(float amount) {
         System.out.println("cash on delivery strat");
     }
 }
